% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_stat_test.R
\name{choose_stat_test}
\alias{choose_stat_test}
\title{Choose Statistical Test Based on Data Characteristics}
\usage{
choose_stat_test(data1, data2 = NULL, paired = FALSE, sigma_known = FALSE)
}
\arguments{
\item{data1}{A numeric vector representing the first dataset.}

\item{data2}{A numeric vector representing the second dataset (optional).}

\item{paired}{A boolean indicating whether the data is paired (default is FALSE).}

\item{sigma_known}{A boolean indicating whether the population variance is known (default is FALSE).}
}
\value{
A string indicating the recommended statistical test.
}
\description{
This function recommends an appropriate statistical test based on data characteristics such as normality, 
variance equality, and whether the data is paired.
}
